ui: restructure com to reuse msg-views between thread and list
