+++ ui/com/user-profile.jsx
@@ -1,8 +1,8 @@
 'use babel'
 import React from 'react'
 import MsgList from './msg-list'
-import Card from './msg-list/card'
-import Oneline from './msg-list/oneline'
+import Card from './msg-view/card'
+import Oneline from './msg-view/oneline'
 import Tabs from './tabs'
 import { VerticalFilledContainer } from './index'
 import { UserInfoHeader, UserInfoFolloweds, UserInfoFollowers, UserInfoFlags } from './user-info'
