+++ ui/less/com/msg-list.less
@@ -1,8 +1,3 @@
-@import "msg-list/oneline.less";
-@import "msg-list/summary.less";
-@import "msg-list/card.less";
-@import "msg-list/notification.less";
-
 .msg-list {
   display: flex;
   .empty-msg {
