+++ ui/less/com/msg-thread.less
@@ -1,0 +1,25 @@
+.msg-thread {
+  &>.toolbar {
+    margin: 0 40px 0 0;
+  }
+  & > div {
+    padding-right: 40px;
+  }
+  .rootlink {
+    margin: 6px;
+  }
+  .composer-container {
+    padding-right: 120px;
+  }
+  .composer {
+    width: 529px;
+    margin: 10px auto;
+    padding-top: 1px; // make background take full space
+    position: relative;
+    left: 36px;
+    box-shadow: none;
+    .composer-audience {
+      display: none;
+    }
+  }
+}
