+++ ui/less/com/msg-view.less
@@ -1,143 +1,4 @@
-.msg-view {
-  display: flex;
-  width: 600px;
-  margin: 10px auto;
-
-  .avatar {
-    flex: 0 0 68px;
-    padding: 2px 0 0 3px;
-    img {
-      width: 60px;
-      margin: 2px;
-      border-radius: 3px;
-    }
-  }
-  .content {
-    .elevation-card;
-    flex: 1;
-    background: #fff;
-  }
-
-  .header {
-    margin: 8px 10px;
-    display: flex;
-    color: #555;
-    cursor: pointer;
-    & > div:first-child {
-      flex: 1;
-    }
-    a {
-      color: #555;
-    }
-  }
-  .body {
-    padding: 0 10px;
-    font-weight: 300;
-    color: @dark-gray3;
-    font-size: 90%;
-    line-height: 1.4;
-
-    .pretty-raw {
-      margin: 16px 0;
-    }
-  }
-  .signallers {
-    padding: 10px;    
-    a {
-      color: #555;
-      font-weight: 300;
-      font-size: 15px;
-    }
-    a.vote {
-      &.selected {
-        color: #1F751F;
-      }
-      &:hover {
-        color: green;
-      }
-    }
-    a.flag {
-      float: right;
-      &.selected, &:hover {
-        color: #BB1111;
-      }
-    }
-  }
-  .signals {
-    background-color: #fafafa;
-    padding: 5px 10px;
-    color: #808080;
-    font-family: "Source Sans Pro";
-    font-weight: 300;
-    font-size: 14px;
-    border-top: 1px solid #eee;
-    a {
-      color: #555;
-    }
-  }
-}
-.msg-view-collapsed {
-  display: flex;
-  width: 600px;
-  margin: 3px auto;
-  cursor: pointer;
-
-  .avatar {
-    flex: 0 0 65px;
-    padding: 2px 6px 0 0px;
-    text-align: right;
-    img {
-      width: 26px;
-      margin: 3px 0 0;
-      border-radius: 3px;
-    }
-  }
-
-  .content {
-    flex: 1;
-    display: flex;
-  }
-
-  .col {
-    padding: 7px 3px 5px;
-    color: #555;
-    a {
-      color: #555;
-    }
-  }
-  .col:nth-child(2) {
-    flex: 1;
-    white-space: pre;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    font-weight: 300;
-    font-family: @content-font;
-    font-size: 90%;
-    line-height: 1.7;
-  }
-  .col:nth-child(3) {
-    color: #555;
-  }
-}
-.msg-view-thread {
-  &>.toolbar {
-    margin: 0 40px 0 0;
-  }
-  .items {
-    padding: 20px 150px 200px 0;
-  }
-  .rootlink {
-    margin: 6px;
-  }
-  .composer {
-    width: 529px;
-    margin: 10px auto;
-    padding-top: 1px; // make background take full space
-    position: relative;
-    left: 36px;
-    box-shadow: none;
-    .composer-audience {
-      display: none;
-    }
-  }
-}
+@import "msg-view/oneline.less";
+@import "msg-view/summary.less";
+@import "msg-view/card.less";
+@import "msg-view/notification.less";
