+++ ui/less/com/msg-view/card.less
@@ -1,4 +1,4 @@
-.msg-list-item.card-post {
+.msg-view.card-post {
   display: flex;
   max-width: 600px;
   margin: 20px auto;
@@ -108,7 +108,7 @@
   }
 }
 
-.msg-list-item.card-muted {
+.msg-view.card-muted {
   display: flex;
   max-width: 600px;
   margin: 20px auto;
