+++ ui/less/com/msg-view/notification.less
@@ -1,4 +1,4 @@
-.msg-list-item.notification {
+.msg-view.notification {
   display: flex;
   max-width: 600px;
   margin: 20px auto;
