+++ ui/less/com/msg-view/oneline.less
@@ -1,4 +1,4 @@
-.msg-list-item.oneline {
+.msg-view.oneline {
   display: flex;
   height: 24px;
   padding: 6px 0 4px;
