+++ ui/less/com/msg-view/summary.less
@@ -1,4 +1,4 @@
-.msg-list-item.summary {
+.msg-view.summary {
   display: flex;
   padding: 10px;
   overflow: hidden;
