+++ ui/less/main.less
@@ -30,6 +30,7 @@
 @import "com/hexagons.less";
 @import "com/suggest-box.less";
 @import "com/msg-list.less";
+@import "com/msg-thread.less";
 @import "com/msg-view.less";
 @import "com/user-list.less";
 @import "com/user-info.less";
