+++ ui/views/bookmarks.jsx
@@ -1,7 +1,7 @@
 'use babel'
 import React from 'react'
 import MsgList from '../com/msg-list'
-import Oneline from '../com/msg-list/oneline'
+import Oneline from '../com/msg-view/oneline'
 import * as HelpCards from '../com/help/cards'
 import app from '../lib/app'
