+++ ui/views/data.jsx
@@ -1,7 +1,7 @@
 'use babel'
 import React from 'react'
 import MsgList from '../com/msg-list'
-import Card from '../com/msg-list/card'
+import Card from '../com/msg-view/card'
 
 export default class Data extends React.Component {
   render() {
