+++ ui/views/newsfeed.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 import pull from 'pull-stream'
 import mlib from 'ssb-msgs'
 import MsgList from '../com/msg-list'
-import Card from '../com/msg-list/card'
+import Card from '../com/msg-view/card'
 import * as HelpCards from '../com/help/cards'
 import WelcomeHelp from '../com/help/welcome'
 import app from '../lib/app'
